Rename k_medoids_.py -> _k_medoids.py
